refactor: split gem avatar content helpers
